Extracting RSolr monkey patch to initializer
